Fix the temp value to stay uint16_t*

+++ ap471Sup/wpnt471.c
@@ -74,8 +74,8 @@ uint16_t wpnt471(struct cblk471 *c_blk, uint16_t port, uint16_t point, uint16_t 
     bpos = 1 << point;
     value <<= point;
 
-    temp = (uint16_t)input_long( c_blk->nHandle, &c_blk->brd_ptr->IORegister[port]);
+    temp = input_long( c_blk->nHandle, &c_blk->brd_ptr->IORegister[port]);
     temp = ( temp & ~bpos ) | (value & 1);
-    output_long( c_blk->nHandle, &c_blk->brd_ptr->IORegister[port], (long)temp);
+    output_long( c_blk->nHandle, &c_blk->brd_ptr->IORegister[port], temp);
     return(0);
 }
